Extracted logic for getting user from cookie into service
