+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/App.tsx
@@ -1,7 +1,7 @@
-import React, { FC } from 'react';
+import React, { FC, useMemo } from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import AuthProvider, { IUserPermissionsType, IUserType } from './hooks/use-auth';
+import AuthProvider from './hooks/use-auth';
 import PageHeader from './layout/header/PageHeader';
 import PageContent from './layout/content/PageContent';
 import PageFooter from './layout/footer/PageFooter';
@@ -23,7 +23,7 @@ import ContestsProvider from './hooks/use-contests';
 import ContestCategoriesProvider from './hooks/use-contest-categories';
 import CurrentContestResultsProvider from './hooks/contests/use-current-contest-results';
 import ContestStrategyFiltersProvider from './hooks/use-contest-strategy-filters';
-import { getCookie } from './utils/cookies';
+import UserCookiesService from './services/user-cookies-service';
 
 interface IProvider {
     Provider: FC,
@@ -63,35 +63,12 @@ const InitProviders = ({ providers, children } : IInitProviderProps) => {
         );
 };
 
-const defaultState = {
-    user: {
-        username: '',
-        isLoggedIn: null,
-        permissions: { canAccessAdministration: false } as IUserPermissionsType,
-    },
-};
-
-const tryGetUserDetailsFromCookie = () => {
-    const loggedInUsername = getCookie('logged_in_username');
-    const canAccessAdministrationCookie = getCookie('can_access_administration');
-    let { permissions } = defaultState.user;
-    let loggedIn = false;
-
-    if (loggedInUsername) {
-        const canAccessAdministration = canAccessAdministrationCookie.length > 0;
-        permissions = { canAccessAdministration } as IUserPermissionsType;
-        loggedIn = true;
-    }
-
-    return {
-        username: loggedInUsername,
-        isLoggedIn: loggedIn,
-        permissions,
-    } as IUserType;
-};
-
 const App = () => {
-    const user = tryGetUserDetailsFromCookie();
+    const userCookiesService = useMemo(
+        () => new UserCookiesService(),
+        [],
+    );
+    const user = userCookiesService.getUser();
     const providers = [
         UrlsProvider,
         ServicesProvider,
